Split layer switcher components
