+++ src/Geopilot.Frontend/src/hooks/useForceUpdate.ts
@@ -1,0 +1,6 @@
+import { useCallback, useState } from "react";
+
+export const useForceUpdate = () => {
+  const [, setTick] = useState(0);
+  return useCallback(() => setTick(tick => tick + 1), []);
+};
